% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_data_cod_mun_INE.R
\docType{data}
\name{pjp_data_cod_mun_INE}
\alias{pjp_data_cod_mun_INE}
\title{Codigos municipales (SPAIN)}
\format{A data frame with 8.136 rows(municipios) y 10 variables
\itemize{
\item INECodMuni:   Codigo INE del municipio  (5 digitos)
\item INECodMuni.n: Nombre del municipio
\item capital_prov:   1 = el municipio es capital de provincia
\item capital_CCAA:   1 = el municipio es capital de CA
\item INECodProv:   Codigo INE de la provincia  (2 digitos)
\item INECodProv.n: Nombre de la provincia
\item INECodCCAA: Código INE de la C.A. (2 digitos)
\item NombreCCAA: Nombre de la C.A.
\item year_first:
\item year_last:
}}
\usage{
pjp_data_cod_mun_INE
}
\description{
Codigos municipales (SPAIN)
}
\examples{
\dontrun{
 cod_mun_INE <- pjp_data_cod_mun_INE
}

}
\keyword{datasets}
