% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_data_cod_prov_INE.R
\docType{data}
\name{pjp_data_cod_prov_INE}
\alias{pjp_data_cod_prov_INE}
\title{Codigos provinciales (SPAIN)}
\format{A data frame with 52 rows (provincias) y 4 variables
\itemize{
\item INECodProv:   Codigo INE de la provincia  (2 digitos)
\item INECodProv.n: Nombre de la provincia
\item INECodCCAA: Código INE de la C.A. (2 digitos)
\item NombreCCAA: Nombre de la C.A.
}}
\usage{
pjp_data_cod_prov_INE
}
\description{
Codigos provinciales (SPAIN)
}
\examples{
\dontrun{
 cod_prov_INE <- pjp_data_cod_prov_INE
}

}
\keyword{datasets}
