% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_data_df_prueba.R
\docType{data}
\name{pjp_data_df_prueba}
\alias{pjp_data_df_prueba}
\title{df de prueba}
\format{A data frame with 20 rows y 6 variables:
\itemize{
\item X_factor: es factor
\item X_integer: integer
\item X_logical: logical
\item X_numeric: double
\item X_text: character
\item X_POSIXct: POSIXct  POSIXt
}}
\usage{
pjp_data_df_prueba
}
\description{
Es sólo un df para probar mis funciones: tiene un factor, integer, double, text, logical y POSIXct
}
\examples{
\dontrun{
 df_prueba <- pjp_data_df_prueba
}

}
\keyword{datasets}
