% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_f_decimales.R
\name{pjp_f_decimales}
\alias{pjp_f_decimales}
\title{redondead un df dejando los decimales q le digas}
\usage{
pjp_f_decimales(df, nn = 2)
}
\arguments{
\item{df}{Un dataframe}

\item{nn}{el nº de decimales (Default = 2)}
}
\value{
Un nuevo \code{df} con sus valores redondeados a \code{nn}
}
\description{
pjp_df_decimales() te retorna un df con los decimales que le digas (default = 2)
pjp_df_decimales function redondea los valores de un dataframe. Le metes un df y redondea sus valores
}
\examples{
pjp_f_decimales(cars, nn = 2)
}
