% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_f_dicc.R
\name{pjp_f_dicc}
\alias{pjp_f_dicc}
\title{returns a df with el diccionario de un df}
\usage{
pjp_f_dicc(df, truncate = TRUE, nn_truncate = 100)
}
\arguments{
\item{df}{(a data frame)}

\item{truncate}{(If you want to truncate the uniques values. The default = TRUE)}

\item{nn_truncate}{(if you truncate, this is the maximun character that you will see. The default = 500)}
}
\value{
Un nuevo \code{df} con los valores unicos del df original
}
\description{
pjp_f_dicc() returns a df con el diccionario del input df
}
\examples{
pjp_f_dicc(iris)
}
