% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_f_round_nice.R
\name{pjp_f_round_nice}
\alias{pjp_f_round_nice}
\title{round numbers nicely}
\usage{
pjp_f_round_nice(..., digits = 2)
}
\arguments{
\item{digits}{(digitos a mostrar, default=2)}
}
\description{
Ya lo he dicho round numbers nicely
This function rounds data in a nice way. See: https://twitter.com/hspter/status/314858331598626816
Thanks Karl Broman!!
}
\examples{
 pjp_f_round_nice(c(-0.0002, 0.0002, 0.5, -0.5, -0.002), digits = 3)
 should return c("0.000", "0.000", "0.500", "-0.500", "-0.002")
}
