% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_f_valores_unicos.R
\name{pjp_f_valores_unicos}
\alias{pjp_f_valores_unicos}
\title{returns a df with the unique values in each colum of athe input df}
\usage{
pjp_f_valores_unicos(df, nn_pjp = 50)
}
\arguments{
\item{df}{(a data frame)}

\item{nn_pjp}{(cuantos valores unicos permites como maximo, default=50)}
}
\description{
my_df_valores_unicos() te retorna un df con los valores(unicos) que existen en cada columna de un df
}
\examples{
pjp_f_valores_unicos(cars)
}
