% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_not_in.R
\name{pjp_not_in}
\alias{pjp_not_in}
\title{mi version de not_in}
\usage{
pjp_not_in(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
Un nuevo vector con los elemntos de x q no estan en y
}
\description{
te dice los valores de x q no estan en y
}
