% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_template_rmd.R
\name{pjp_template_rmd}
\alias{pjp_template_rmd}
\title{call and open la plantilla para crear un archivo .Rmd}
\usage{
pjp_template_rmd(name = "titulito", open = interactive())
}
\arguments{
\item{name}{name of your analysis file}

\item{open}{si estas en interactivo o no}
}
\description{
Uuar la plantilla para crear un archivo .Rmd
}
\examples{
pjp_template_rmd()
}
