% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_unique_values.R
\name{pjp_unique_values}
\alias{pjp_unique_values}
\title{returns a df with the unique values in each colum of the input df}
\usage{
pjp_unique_values(df, truncate = FALSE, nn_truncate = 500)
}
\arguments{
\item{df}{(a data frame)}

\item{truncate}{(If you want to truncate the uniques values. The default = TRUE)}

\item{nn_truncate}{(if you truncate, this is the maximun character that you will see. The default = 500)}
}
\value{
Un nuevo \code{df} con los valores unicos del df original
}
\description{
pjp_unique_values() returns a df with the unique values that exits in each column of the input df
}
\examples{
pjp_unique_values(iris)
}
