% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjpv2020.01-package.R
\docType{package}
\name{pjpv2020.01-package}
\alias{pjpv2020.01}
\alias{pjpv2020.01-package}
\title{pjpv2020.01: Paquete de uso personal}
\description{
Paquete para uso personal con algunas
    funciones y conjuntos de datos
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/perezp44/pjpv2020.01}
  \item Report bugs at \url{https://github.com/perezp44/pjpv2020.01/issues}
}

}
\author{
\strong{Maintainer}: Pedro Perez \email{pedro.j.perez@uv.es}

}
\keyword{internal}
