% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pjp_maps.R
\name{theme_pjp_maps}
\alias{theme_pjp_maps}
\title{creo un theme apropiado para mapas}
\usage{
theme_pjp_maps()
}
\value{

}
\description{
theme_pjp_maps() crea un theme apropiado para mapas
}
\details{
\link{ggplot2} apropiado para mapas
}
\examples{
theme_pjp_maps()
}
