% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp_template_bitacora.R
\name{pjp_template_bitacora}
\alias{pjp_template_bitacora}
\title{call and open la plantilla para crear una bitacora}
\usage{
pjp_template_bitacora(name = NULL, open = interactive(), ...)
}
\arguments{
\item{name}{name of your TidyTuesday analysis file}
}
\description{
Uuar la plantilla para bitacoras y empezar a documentar el Rproject
}
\examples{
pjp_template_bitacora()
}
